[Rocky Only] Fix upgrade_steps_tasks playbook generation.
In the backport [0] it was removed the loop that iterated
over each role to create the upgrade_steps_tasks playbook
and included the generation as part of the previous
secion (pre_upgrade_rolling_steps_playbook) loop.
This causes the upgrade to stop working as upgrade_steps_
tasks will be build for only one role (the last we iterated
over in the previous loop) and running the upgrade for any
other role will end up in all the upgrade_tasks being
skipped.

[0] - I4ee0110a6c2b9466d81e37e5df27f5f81a6eceb5
Change-Id: Iff7308b544f73fa698b4e532ca64a3ace19905ef
Closes-Bug: #1848789