[QueensOnly] Update ovs to 2.11 without network loss
During the minor update from ovs2.9 to ovs2.11,
handled the package name changes to ensure that
ovs is not restarted. This change have been handled
already on the upgrade path (queens -> rocky). Handle
the same in the minor update (update_steps).

This change is not required in master or other stable
versions, because all other versions are already
migrated to ovs2.11 package and the upgrade path
is handled.

Change-Id: Id563806777679998412a53c693b02152df1e800e