Add the ability to configure ovn-remote-probe-interval
This parameter sets inactive probe interval of the JSON
session from ovn-metadata to the OVN SB database.
By default this it is 5s which not be sufficient in
loaded systems or during high control-plane activity spikes,
leading to unnecessary reconnections to OVSDB server.
Now it is extended by default to 1 min and it
is configurable.

Depends-On: https://review.opendev.org/#/c/695569
Change-Id: Ied8f7939a32bfcc6ae5563b625b438a1492a5e2c