Adds ceph-systemd tag in external_upgrade_tasks for systemd units
When upgrading the host operating system we'll also replace docker
with podman; ceph-ansible provides for a special playbook meant
to refresh the systemd units to work with podman.

This change adds a task which on upgrade will execute the special
playbook. It requires ceph_ansible_limit to be set by the user as
that gets translated into a --limit option for the playbook execution
so that it runs only against a specific node undergoing upgrade
of the host operating system.

Change-Id: I99bc5342360e307a9ee499cde5f2a0b5d78170f8
Depends-On: I4538d64ed49b4a6e4670d1e1d050fc5443e2edb0
(cherry picked from commit e8048707f7a56e52d8664ece8a4adbf958c0001e)