Add "radvd_user" config parameter to Neutron L3 container
The username passed to radvd, used to drop root privileges and change
user ID to username and group ID to the primary group of username. If no
user specified, the user executing the L3 agent will be passed. If "root"
specified, because radvd is spawned as root, no "username" parameter
will be passed

Depends-On: https://review.opendev.org/#/c/691846/
Change-Id: Ie9f16f3117e6531f4ef1e4973e61c72dbf904b8a
Closes-Bug: #1844688