Install and start Rsyslog on the Undercloud & Standalone
Create a new Rsyslog service that is deployed on the host (not in a
container) and with Ansible.

Make it so it's deployed by default on Undercloud & Standalone setups.
Also move the tasks that configure rsyslogd for HAproxy & Swift to be
executed after the host prep tasks (using deploy step tasks).

Change-Id: I027c64aefcc4715da17836a5cf0141152cf146aa
Closes-Bug: #1850562
(cherry picked from commit 5ff628f1cfb72a88e7414370e0f9083a206eb710)