Set ServerDeletionPolicy during upgrade prepare and converge
This change will always set ServerDeletionPolicy to 'retain' during
upgrade prepare, and back to the default 'delete' during the converge.

For a normal upgrade this will have no effect since scale-down is not
recommended during an upgrade.

For a transition from nova based undercloud to nova-less, this allows
the following steps to work:
1. operator builds a valid deployed-server environment file by running
   'openstack overcloud node extract provisioned'[1]
2. running 'openstack overcloud upgrade prepare' will prevent the next
   stack update from deleting any nova servers
3. running 'openstack overcloud upgrade converge' with the environment
   from 1. will replace nova heat resources with deployed server
   resources which are managed directly with the provision/unprovision
   commands

This approach is an alternative to preventing the heat nova resource
from ever deleting a nova server after the undercloud upgrade[2]

[1] https://review.opendev.org/#/c/674141/
[2] https://review.opendev.org/#/c/675498/
Blueprint: bp/nova-less-deploy

Change-Id: I0bca04cd4883a1bf61b06a0828e323bc388a95be