Move bootstrap password to an environment var
Currently if you try to use a space in the keystone password, the
bootstrap process fails due to the password being evaulated as command
line arguments. We can work around this by using an environment var to
pass the password to the script that needs to be run.

Conflicts:
        deployment/keystone/keystone-container-puppet.yaml

Change-Id: I03754206781bc8f62d143b2c541f594ca3207a0f
Closes-Bug: #1811005
(cherry picked from commit 068527d139b4b521782873fdd97ac46cb10b0885)