zuul: remove "files" for openstack-tox-tht
openstack-tox-tht is a dependency of functional testing jobs, which can
be problematic now because it only runs if some files are touched.

If e.g. common/container_puppet.py is the only file that is changed,
zuul will complain because the dependency graph is broken.

Let's relax openstack-tox-tht to run all the times; and stay a
dependency of other functional testing jobs (which brings value).

Change-Id: I3f05ca26064a2897a724876033a5cc4a602c38a2