Define ovn_dbs_image fact for upgrades.
The ovn upgrade_tasks make use of a block which belongs
to the update_tasks. In that block, the ovs_dbs_image
variable is being referenced. However, the fact is set
in a step before the block invoked.

This patch sets ovn_dbs_image fact in step0 for upgrade
tasks and fixes a small typo with the bundle.

Change-Id: I1d24dce0ff9c88380e8769593d619a9e4e944e40
Closes-Bug: #1851669
(cherry picked from commit 460d4771a6be1b7aebc8c86263149e2835cad654)