Split QDR environment from collectd
This patch splits QDR configuration from shared environment files which enables
flexibility to either run in edge mode only or in mesh mode.

It also adds separate environment file for ceilometer and dynamic listener
configuration for storage nodes for cases when storage nodes are not connected
to internal network (which is a case for CephStorage nodes by default).

Change-Id: I6b5af0b5afcdec75fe239385922dc84fbbb8288c