ovn-db standalone: Create ovndbs before starting the OVN ovsdb servers
The kolla commit [1] deleted the extend_start script for ovn-nb-db-server/ovn-sb-db-server
kolla images. This script created the ovn dbs from the schema. After this commit
tripleo CI is broken as the ovn dbs are not present.

To fix this issue, this patch first creates OVN dbs before starting the OVN ovsdb-servers.

[1] -  I31a372d801dace92950cd85b29b308431f78bbdb
Closes-bug: #1852682
Change-Id: I9640529455e1d14c98c3b53993a71d77de8d7a25
Signed-off-by: Numan Siddique <nusiddiq@redhat.com>