Add STF environment
Add yaml to enable the STF client side. The only manual configuration
necessary should be pointing the host to the STF server.

Remove tcpconns from the collectd default plugin list, as
there are no users of it, and it should not be in the default
plugin list when STF is enabled.

More discussion about settings at
https://github.com/infrawatch/telemetry-framework/pull/100

Note about the plugins:

- virt plugin is already automatically loaded on each role where
  nova-compute is deployed. The hieradata is always set in
  eployment/metrics/collectd-container-puppet but it's fine because the
  plugin isn't always enabled. No need to add extra complexity with
  parameters.
- cpu, df, load, intel_rdt, connectivity, procevent and ipmi are
  automatically loaded if STF is enabled.

Related: https://review.opendev.org/#/c/710376/

Signed-off-by: Ryan McCabe <rmccabe@redhat.com>
Signed-off-by: pleimer <pfbleimer@gmail.com>
Signed-off-by: Leif Madsen <lmadsen@redhat.com>
Co-Authored-By: Emilien Macchi <emilien@redhat.com>

Change-Id: Ifb1923bff9a344985b3e4de04ade4abd4916d2d6