Make sure we apply all deploy step-0 during update.
That step include, in particular, the puppet-container.py/sh creation
file plus some other task that should happen very early during the
deploy process.

During update if a change in puppet-container.py happen in the
template we have to trigger those steps to get the changes.  This is
especially critical because we run puppet container configuration
during the update stage.

Closes-Bug: #1853156
Change-Id: I26406da82c584dc5093c17ad26f263057a5cbcaa