Update scale down to fix regression
This change updates our scale down process to ensure that we're
searching for and matching only the one host that intended to
be remove during the task execution. The process will now use
the `ansible_fqdn` fact as the search criteria and ensure we're
only ever interacting with one known host.

* If the search criteria returns more than one value the task
  will fail indicating why it failed and return data on the
  failure.
* If the search criteria returns no values, it will be assumed
  that the compute service is already disabled and has been
  removed.

Resolves: rhbz#1495489
Closes-bug: 1853644
Change-Id: I2f684191b146e0f6a28b98d7c0080e6eab14f2d9
Signed-off-by: Kevin Carter <kecarter@redhat.com>