[Stein Only] Fix bad configuration path in nova-placement wait
Change I6fcbc5cb5d4f3cbb618d9661d2a36c868e18b3d6 implements a wait for
the nova-placement API to be available. The backport did not adjust the
configuration path for nova-placement to be compatible with a Stein
deployment. This patch sets the correct configuration path for Stein only.

Closes-Bug: 1853504
Related-Bug: 1842687
Change-Id: Ibfaaa5b114dd07155d78f9d2d84a4c4a14c22ac8