Drop file test before removal
If the file is already missing (e.g. has been removed already), the test
-f will return 1 which fails the container-puppet.sh execution. The goal
here is to just make sure the file is missing, so let's just try to
removing it with only an rm -f. This is successful if the file is
already missing so we don't need to check the file existance.

Change-Id: I41bbff148267de6514bf5b88c344d18cb894ae01
Closes-Bug: #1853183
(cherry picked from commit 3ade96f7688e890140bf80cc228d3fc057ef80dc)