Don't fail hard when facter cache doesn't work correctly
Since the facter cache is supposed to be a help but not necessarily a
hard requirement, let's make sure the tasks don't fail if something goes
wrong. We should output the error if the facter execution fails, but we
don't want to always output it.

Change-Id: I58d42c74295f031b260c0b67dbc12336fc31db1f
Related-Bug: #1852606
(cherry picked from 79f41d1d41a23b8b851410a912fd352c5e6fac32)