Create a new parameter for the HAProxy external network
In the standalone config, there is no external network defined.
This leaves the hieradata cloud_name_external undefined, resulting in
an error when we are using the haproxy-public-tls-certmonger template
to create new public certs through haproxy.

Using the PublicNetwork allows us to get the right network for haproxy
in all cases to specify the network and fqdn to use for public certmonger
certificates.

Co-Authored-By: Grzegorz Grasza <xek@redhat.com>
Change-Id: I8778402bbb7a670c4aa95671c6017dff356238d4