Force redis::ulimit to be an integer
The upstream change at:
https://github.com/voxpupuli/puppet-redis/commit/502ff38c8c035fa143e32dd4ee61b3ec68db461c

made redis::ulimit an integer and will now error out if passed a string.
Historically in THT we passed RedisFDLimit directly to the redis::ulimit
hiera key as a string. Let's just change RedisFDLimit to a number. The
only affected people will be those that A) tweaked this parameter and
B) actually wrapped the param into quotes so to force it as a string.
Given that this number is likely non-significance, we choose this route
and we avoid an extra YAQL invocation to do a conversion in THT.

Change-Id: I70361801c71738f2d95930aea99914316d9f0d11
Closes-Bug: #1854136