Deploy /etc/openstack/clouds.yaml with Ansible
Replace the python script that was run on post-config, by an Ansible
task running on the host where Keystone is running.

It'll be useful later when using OpenStackSDK to have access to the
credentials during the deployment and not having to wait the far end.

It's also reducing the Heat resources.

Change-Id: I585abc3e6a3b9b8ae9183e0b5170df2e39301e17
(cherry picked from commit 25a2045e6205d07f38cc6a065f9ad67b86ebe43d)