Move 'Ensure network service is enabled' after os-net-config has run
This has been observed downstream via https://bugzilla.redhat.com/show_bug.cgi?id=1777529
The scenario is as follows:
1) Undercloud VM is configured with a bunch of interfaces in dhcp
2) One NIC (eth0 in this case) is on a network without a dhcp server
3) Said NIC is configured in undercloud.conf to be part of the
   br-ctlplane bridge

In the above scenario the undercloud install will fail with the
following error:
 "Unable to start service network: Job for network.service failed because the control process exited with error code...."

The reason is that eth0 times out when asking for dhcp and so the whole
network.service start command fails.

We can just move the network service enablement *after* os-net-config
has run. This fixes this scenario (I just tested it) and it still
covers for enabling the network service in order to maintain network
connectivity at reboot, which was the reason for adding this code in the
first place anyway.

Change-Id: I0d13c9ee2490aa765d546961c9a1fc14e931c0c7