ovn_dbs upgrade-with-os work
This adds the missing steps for the upgrades-with-os blueprint for
the ovn_dbs service. The ovn_dbs pacemaker service was not taken care of
in the initial work in https://github.com/openstack/tripleo-heat-templates/commit/cb8caaae4381e6a3de81a517dc7c6b23d272e024
This commit tries to fix that.

Closes-Bug: #1854519

Tested-by: Jose Luis Franco Arza <jfrancoa@redhat.com>

Depends-On: I7bff96c0b0238e4aeaa33e0ec343a75e08b6a2f7
Change-Id: Id90754ad2506ee52d2a2cb939ccbf2d84e792cff