New Parameter NovaCronArchiveDeleteAllCells and NovaCronArchiveDeleteRowsAge
Introduces two new parameters to configure the archive deleted
instances cron job.
1) NovaCronArchiveDeleteAllCells
To make sure deleted instances get archived also from the cell0
in a single cell deployment and also in additional cell databases
in case of a multi cell deployment.

2) NovaCronArchiveDeleteRowsAge
--before is required to prevent the orphaning of libvirt guests
if/when nova-compute is down when a db archive cron job fires.

This change also modifies
1) the default from 100 to 1000 for NovaCronArchiveDeleteRowsMaxRows
to match the default from the nova-manage command instead the default
of 100 from the puppet-nova parameter.

2) changes the default for NovaCronPurgeShadowTablesAllCells from
false to true also the nova-manage db purge command needs to run
for all cells instead of only the default cell.

Depends-On: https://review.opendev.org/696900
Depends-On: https://review.opendev.org/697299

Change-Id: I91cb1e16f65b23117235d4eac76f03748b47e926