Re-enable "service_names" hieradata
"service_names" was a useful hieradata which listed the services enabled
for a specific roles vs "enabled_services" which are for all the
services enabled in the cloud, no matter the role.

This is re-added for backward compatibility.

Depends-On: I75b1112089a66cf5db0a2fd651bb24428cf861fd
Related-Bug: #1855138
Change-Id: I7339b8791817bdaffa65c928d424796114efdf57
(cherry picked from commit b5eff6783050a0449d4da1027e38dcee17672c66)