Ensure rabbitmq deployment container is restarted at every update
In [1] we lost the ability to restart deployment containers such as
rabbitmq_bootstrap and rabbitmq_init_bundle every time a stack update
or minor update is run. A fix for it [2] was attempted but was only
partial so not working.

Make sure we reintroduce variable TRIPLEO_DEPLOY_IDENTIFIER in all
the rabbitmq deployment containers (HA and non-HA) to make sure they
get restarted as expected at every update.

[1] I894f339cdf03bc2a93c588f826f738b0b851a3ad
[2] I139c2c276c8f532d37fceafe06f9547048e76906

Change-Id: I131a905b62cd990deee1ae6aa81ab2560c98a7c8
Closes-Bug: #1855732