Use async tasks for long running common tasks
Use async with poll: 0 for the long running common tasks including:

- puppet host configuration
- container-puppet generate config
- container starting
- container-puppet bootstrap tasks

Executing the tasks in this manner and then polling for the results
causes all of the tasks to be started in parallel across all the nodes,
regardless of the number of ansible forks configured.

The behavior will be that ansible will start the task on the first count
of nodes that matches the configured forks value. Since poll:0, ansible
will immediately move on to the next batch, etc. Effectively, all the
tasks are started in parallel just as quickly as ansible can start them.

The polling tasks (async_status) will then execute in parallel up to the
configured number of forks. As tasks start to finish, ansible moves on
to checking the status of the next batch (again, up to the configured
number of forks). Since most nodes will configure around the same time,
the polling tasks finish roughly at the same time (except for
differences in roles, such as controllers taking much longer).

This behavior results in a signifcant performance improvement at scale,
or when deploying any number of nodes greater than the configured forks
value. Instead of waiting for the first batch to fully complete, all the
nodes are started in parallel.

For example, if puppet host configuration usually takes 5 minutes per
node, and there are 100 nodes, with 25 forks, previously this would have
taken 20 minutes. With this patch, it would now take closer to 5 minutes
for all 100 nodes, plus some overhead for polling.

This more closely matches the behavior previously used with Heat, when
all the nodes were operating in "pull" mode in parallel.

Change-Id: I0c651d127cd2bb179f7592a0519a5fd5064faeb3
(cherry picked from commit 106ce3267fd1471cee184f697f7cb19c695ab36c)