Add ntpdate to ensure hwclock sync
The hwclock is not kept insync with the system unless the ntpdate
service is configured to sync the service. Since we previously didn't
enable the ntpdate service, we are configuring it to ensure hwclock sync
and that it is started if ntpd is being used.  Chrony has rtcsync by
default.  ntpd was dropped in Train so this is necessary for <= Stein

Change-Id: I51fb67e0bfdd03fc74fe242965648ca52c4338d7
Closes-Bug: #1856259