Force facts cache refreshing before upgrade.
When upgrading from Rocky to Stein, an upgrade of the operating system is
performed. This upgrade from RHEL7 to RHEL8 implies the removal of the
default /usr/bin/python binary. As the facts cache is enabled, Ansible's
strategy does not consider to upgrade facts and therefore we try to run the
ansible playbook using the old python binary when running the upgrade.
This fails with the error: /usr/bin/python: No such file or directory.

This patch makes use of the setup task in combination with gather_facts
false, to ensure that the facts are gathered and refreshed for the
Overcloud nodes. This way, we make sure that we are using the right
python binary. As during scale, a similar situation is occuring, this
patch adds the same logic in scale_playbook.

Closes-Bug: #1856313
Change-Id: I87974e88c38b42e90bc3cd801fcf1deaf268720c
(cherry picked from commit 14db20baee3529b223a87e9f73279c6034a47427)