deploy-steps: use lookup() to get pre|post_deployments_role variables
Note: this patch is only for Stein and its previous versions.
Train and master are fine.

The syntax in place didn't allow to have an empty default for these 2
vars; which can happen in Queens if config-download is disabled.

Using the lookup module allows to set a default.

Change-Id: I38c658e0b0f2e334bd529fe66f937465036cc205
Closes-Bug: #1856314
(cherry picked from commit 44d2502383bc539287cf42ca488773242ad6be9f)