Create a dedicated log file for healthchecks for collectd
This is a first step in order to get rid of the journalctl/systemctl
calls from within collectd container.

If CollectdEnableSensubility is set to true, the configuration will be
added, and if set to false, it will remove it. The latter is in order to
ensure we don't risk any disk space issue in case the operator wants to
remove this feature.

It depends on two patches that actually enable the SyslogIdentifier
configuration in the healthcheck units.

Change-Id: If4414c6ceced9200f9b0dc97be5043d7b2684ae1
Closes-Bug: #1856575
Depends-On: https://review.opendev.org/701186
(cherry picked from commit 122929c21e33f13a1b3dd146076a6c00aea86637)