Create a dedicated log file for healthchecks for collectd
This is a first step in order to get rid of the journalctl/systemctl
calls from within collectd container.

If CollectdEnableSensubility is set to true, the configuration will be
added, and if set to false, it will remove it. The latter is in order to
ensure we don't risk any disk space issue in case the operator wants to
remove this feature.

It depends on two patches that actually enable the SyslogIdentifier
configuration in the healthcheck units.

Change-Id: If4414c6ceced9200f9b0dc97be5043d7b2684ae1
Closes-Bug: #1856575
Depends-On: Icdc5caf4cedc46291a807c39c0a31c74955a4a74
Depends-On: I6861baa287f2a8288b87be26aacecbcc061cd96f