Create /var/log/journal directory during step-0
While I93dcc57aff63b91dab475b0c114b278324434e41 did the right thing, it
didn't do it in the right place.

This patch adds the creation without removing the misplaced code, so
that we can do an easy backport to Train.
A follow-up will be the actual revert of the other change.

Change-Id: Id937389e4eda5c1fc4634ab14695390a09300468
Closes-Bug: #1856278