mysql: move firewall_rules to mysql-container-puppet
The firewall_rules need to be in the service template and can't be in
the base, that is imported.
The current YAQL is looking for firewall_rules in the role_data:
$.data.role_data, []).where($ != null).select($.get('firewall_rules')

So moving the firewall_rules from mysql-base to mysql-container-puppet
will allow the data to be found by the query and the firewall rules to
be applied.

Change-Id: I6183cdf63ea628cc86742d56b5e2cc0ec5e3aab9
Closes-Bug: #1856864