Correct invalid jinja set
This change corrects an issue with the jinja expression at the
top of the file which is resulting in a stacktrace when the
set attempts to update an object of the same name, which results
in a none type.

This change also updates the jinja expression at the top of the
file making it more legible.

Change-Id: Ia3334c5d8a20bb3c0bc721a266a09b6c01fe7c4d
Closes-Bug: #1856918
Signed-off-by: Kevin Carter <kecarter@redhat.com>