Introduce CephExtraKeys
Introduce the CephExtraKeys parameter which may be used
have the newly deployed Ceph cluster create additional
keys which may be used by clients to access the cluster.

This parameter will useful for the multiple-external-ceph
blueprint because an operator would use CephExtraKeys to
create a cluster with a new key and that same key could
then be passed to a subsequent deployment to access the
same Ceph cluster. The alternative is to require the
operator to create this key manually between doing split
control plane deployments.

Change-Id: Ic47c2ad47e15e7e7bf56b300517e333f7ebd4013
Implements: blueprint multiple-external-ceph