HA: Fix the cluster common tag behaviour with podman
Since [1,2], HA container image name used in pacemaker are
look like container-common-tag/<servicename>:pcmklatest.

Unlike docker, podman prepends a host information in
front of this tag, which confuses the podman resource
agent in pacemaker.

Make the first part of the tag look like a fqdn so that
podman doesn't change it.

Closes-Bug: #1858648

[1] Id369154d147cd5cf0a6f997bf806084fc7580e01
[2] I7a63e8e2d9457c5025f3d70aeed6922e24958049

Change-Id: I07d72b87225dbadbc4df46564452ccb232593219