mysql: refactor upgrade tasks to work without paunch
We are removing Paunch, and also Podman is the only container cli
supported when it comes to upgrades, so this patch simplifies a bit how
we deal with the upgrade of MySQL container.

This submission also adds the path for the ansible modules
installed from tripleo-ansible in the tht-check job.

Change-Id: I8b01ac05bae6544119769e50501bf705b22081f3