horizon: put plugins toggles in quotes
Without quotes, it seems like Ansible "from_yaml" filter will convert
yes to True and no to False; which is problematic in the case of Kolla
image for Horizon where the extend_start script checks for yes/no and
not the actual booleans.

Putting them between quotes seems to solve the issue.

Change-Id: Ie08726c0916c55c3d4c315b63ee341196cc2e70e
Closes-Bug: #1859491