clustercheck: use fqdn instead of ip for bind address
Clustercheck service is accessible via port 9200 on the Mysql network
(e.g. internal API). The bind address specified in the config currently
uses IP address. This breaks the OpenStack healthcheck because curl
refuses to fetch a URI with unquoted IPv6 address.

Use an fqdn as the bind address in the config to make sure the service
is reachable no matter the type of network used.

Change-Id: Ia41bf10ea65edd51cafee5416697628f972c693e
Closes-Bug: #1860124