[TRAIN ONLY] Check for correct file during --skip-deploy-identifier handling.
At some point, the file names that are checked for existence when
determining to apply the logic for --skip-deploy-identifier must have
changed.

Since the file check was not updated, it was checking for a file that
would never exist, and so the ansible tasks were always run (or re-run).

Effectively, --skip-deploy-identifier was broken and it had no effect.

This patch updates the file check to use the correct name, and also
checks the right file per step. It only applies to Train since on
master, these files were switched to write out per container files in a
directory, and the fix was addressed at that time.

Change-Id: Ia01a10c488d22e38cd86b0117e9c048ea5cc3298
Closes-Bug: #1860566