Add DeployIdentifier to extra config containers
Certain config containers might need to be replaced and re-run
regardless of whether configuration changes on update and upgrade.
Adding the DeployIdentifier to the env will ensure that they are.

Conflicts:
  deployment/mistral/mistral-api-container-puppet.yaml

Note: this cherry-pick is missing changes to
ovn-metadata-container-puppet.yaml that  were made in the original
patch because the systemd-tripleo-wrappers was backported, making this
particular change unnecessary.

Change-Id: I150212ebac3fed471ffb4e7ed7b6eb6c7af3fad9
Closes-Bug: #1860571
(cherry picked from commit 714e1b5d3120b1babfec01c15bf508cc9241ceb2)