Dynamically include container-puppet tasks
The container-puppet tasks only need to be run if tasks actually exist, which
is already being checked on the ansible control node.

A "when" statement is then applied to the set of tasks necessary to run
the container-puppet tasks, when the tasks are actually defined.

This patch moves that set of tasks to a separate tasks file and uses a
dynamic include.  This results in less tasks being skipped, which can
save several minutes at scale. This results in 3 less tasks that need to
be skipped at steps 1-5, which equates to 15 tasks overall, when no
container-puppet tasks actually exist.

When container-puppet tasks do exist, all the tasks will be executed as
necessary.

Change-Id: Ifad32bf79942cde58295fd9aae7e23e2f62c1ae2
(cherry picked from commit 1602d68daabecb420efa43188fc517257aa85e0a)