Update ro excludes
When container-puppet.sh (or docker-puppet.sh) runs, we always bind
mount in certain directories and files. However if these files change
while we're running the puppet execution externally on the host, they
end up being copied to our puppet-generated path. This path is copied
into the container when kolla starts up. Because we re-mount these same
paths in the end container as read-only, kolla will try and update the
read-only mount which errors. This change updates our rsync excludes to
include:

 * /etc/pki/ca-trust/extracted
 * /etc/pki/ca-trust/source/anchors
 * /etc/pki/tls/certs/ca-bundle.crt
 * /etc/pki/tls/certs/ca-bundle.trust.crt
 * /etc/pki/tls/cert.pem
 * /etc/hosts
 * /etc/localtime

Change-Id: I858a7b8745688c6271b01795460ab4adbf3f2be5
Closes-Bug: #1860607