Increase concurrency when starting containers
When Paunch is disabled, we can make the container startup concurrent.
The default was set to 2 to avoid podman locks issues but it seems
that we have now better results when concurrency is higher.

Let's set it to 5 for now.

Change-Id: I707b2f66eb947d64ef6a21238dc26c81575be2c7
(cherry picked from commit 7bbee2d86ff4b78bf29992da2e5d53ab27a5a983)